+++ NEWS.md
@@ -2,6 +2,9 @@
 * Cleaned up code in getEOOPolygon.
 * gridUncertainty functions now have default min.percent.rule set to FALSE.
 * gridUncertainty now uses while, eliminating need for max.rounds parameter.
+* gridUncertaintyBase now includes two new parameters: restriction and
+min.grids.shift
+* Added gridUncertaintyRestricted
 
 # redlistr 1.0.0
 * redlistr is now on CRAN!